+++ .storybook/storybook.scss
@@ -2,6 +2,7 @@
 
 @use "../src/nationalarchives/tools/colour";
 @use "../src/nationalarchives/tools/media";
+@use "../src/nationalarchives/tools/typography";
 
 @import "https://use.typekit.net/hkj3kuz.css";
 
@@ -149,7 +150,10 @@
 }
 
 .tna-colour-contrast-demo {
+  width: 0;
+  max-width: calc(100% - 1px);
   display: table;
+  table-layout: auto;
 
   border-top: rgb(0 0 0/0.25) 1px solid;
   border-left: rgb(0 0 0/0.25) 1px solid;
@@ -163,6 +167,8 @@
     padding: 1rem;
 
     display: table-cell;
+
+    @include typography.main-font-weight-bold;
 
     border-right: rgb(0 0 0/0.25) 1px solid;
     border-bottom: rgb(0 0 0/0.25) 1px solid;
@@ -182,6 +188,12 @@
     border-right: rgb(0 0 0/0.25) 1px solid;
     border-bottom: rgb(0 0 0/0.25) 1px solid;
 
+    &:first-child {
+      padding: 1rem;
+
+      @include typography.relative-font-size(14);
+    }
+
     &-content {
       height: 100%;
       padding: 1rem;
@@ -196,9 +208,24 @@
 
   .tna-template {
     min-width: 15.25rem;
+
+    p:not(:first-child) {
+      margin-top: 0.25rem;
+    }
   }
 
   .tna-chip {
     margin: 1rem 0;
+  }
+
+  .tna-button-group {
+    margin-top: 1rem;
+
+    gap: 0.5rem;
+    flex-wrap: nowrap;
+  }
+
+  .tna-pagination {
+    margin-top: 1rem;
   }
 }
