+++ src/nationalarchives/utilities/_typography.scss
@@ -143,6 +143,14 @@ small {
   @include colour.accent;
 
   @include colour.colour-border("accent-background", 0.125rem);
+
+  .tna-background-accent &:not(#{&}--plain) {
+    @include colour.plain;
+
+    // background-color: #f0a !important;
+
+    @include colour.colour-border("background");
+  }
 }
 
 .tna-chip {
